Added two kind of gesture recognizers: tap and swipe. You can choose tap, swipe or both calling addGesture method with the appropriate type..
The Tap gesture localize the touch and if it's on the left part the slideshow slides to previous image, otherwise to the next image.
The Swipe gesture handle the next and previous changing the speed of the animation for a better user experience, it's suggested to use the slide animation if you want to enable this to have a uiscrollview like experience.